# Improve the memory structure using Link Lists